pyproject.toml
[project]
name = "pearscarf"
dynamic = ["version"]
description = "Operational infrastructure that grows itself"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic",
    "openai",
    "click",
    "discord.py",
    "google-api-python-client",
    "google-auth-oauthlib",
    "httpx",
    "langsmith",
    "neo4j",
    "psycopg[binary]",
    "psycopg-pool",
    "qdrant-client",
    "python-dotenv",
    "sentence-transformers",
    "torch",
    "fastmcp",
]

[project.scripts]
pearscarf = "pearscarf.interface.cli:cli"
psc = "pearscarf.interface.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "pearscarf/__init__.py"

[tool.pyright]
extraPaths = ["experts"]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
    "experts/*/eval",
    "lab",
    "data",
]

[tool.ruff.lint]
# E: pycodestyle errors, F: pyflakes, W: pycodestyle warnings,
# I: isort, UP: pyupgrade, B: bugbear
select = ["E", "F", "W", "I", "UP", "B"]
# Pearscarf uses `from __future__ import annotations` extensively + existing
# style has some legitimately long lines in prompts / Cypher queries.
ignore = [
    "E501",  # line too long — handled by formatter; some lines intentionally long
    "B008",  # function call in default argument — common Click / dataclass pattern
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]  # scripts do env setup before imports
"experts/*/*_ingest.py" = ["E402"]  # same pattern in expert ingester entry points

# Pin `torch` to the CPU-only wheel on linux. `sentence-transformers` pulls
# torch transitively; on linux x86_64 its default wheel is the CUDA build,
# which drags in ~4 GB of nvidia + triton libraries we never run. macOS
# torch is already CPU-only by default — no override needed there.
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
